+++ packages/launchpad/src/pubkey.ts
@@ -1,5 +1,5 @@
 import { Bech32, fromBase64, fromHex, toBase64, toHex } from "@cosmjs/encoding";
-import equal from "fast-deep-equal";
+import { arrayContentEquals } from "@cosmjs/utils";
 
 import { PubKey, pubkeyType } from "./types";
 
@@ -21,12 +21,13 @@ const pubkeyAminoPrefixEd25519 = fromHex("1624de6420");
 const pubkeyAminoPrefixSr25519 = fromHex("0dfb1005");
 const pubkeyAminoPrefixLength = pubkeyAminoPrefixSecp256k1.length;
 
-export function decodeBech32Pubkey(bechEncoded: string): PubKey {
-  const { data } = Bech32.decode(bechEncoded);
-
+/**
+ * Decodes a pubkey in the Amino binary format to a type/value object.
+ */
+export function decodeAminoPubkey(data: Uint8Array): PubKey {
   const aminoPrefix = data.slice(0, pubkeyAminoPrefixLength);
   const rest = data.slice(pubkeyAminoPrefixLength);
-  if (equal(aminoPrefix, pubkeyAminoPrefixSecp256k1)) {
+  if (arrayContentEquals(aminoPrefix, pubkeyAminoPrefixSecp256k1)) {
     if (rest.length !== 33) {
       throw new Error("Invalid rest data length. Expected 33 bytes (compressed secp256k1 pubkey).");
     }
@@ -34,7 +35,7 @@ export function decodeBech32Pubkey(bechEncoded: string): PubKey {
       type: pubkeyType.secp256k1,
       value: toBase64(rest),
     };
-  } else if (equal(aminoPrefix, pubkeyAminoPrefixEd25519)) {
+  } else if (arrayContentEquals(aminoPrefix, pubkeyAminoPrefixEd25519)) {
     if (rest.length !== 32) {
       throw new Error("Invalid rest data length. Expected 32 bytes (Ed25519 pubkey).");
     }
@@ -42,7 +43,7 @@ export function decodeBech32Pubkey(bechEncoded: string): PubKey {
       type: pubkeyType.ed25519,
       value: toBase64(rest),
     };
-  } else if (equal(aminoPrefix, pubkeyAminoPrefixSr25519)) {
+  } else if (arrayContentEquals(aminoPrefix, pubkeyAminoPrefixSr25519)) {
     if (rest.length !== 32) {
       throw new Error("Invalid rest data length. Expected 32 bytes (Sr25519 pubkey).");
     }
@@ -55,17 +56,42 @@ export function decodeBech32Pubkey(bechEncoded: string): PubKey {
   }
 }
 
-export function encodeBech32Pubkey(pubkey: PubKey, prefix: string): string {
+/**
+ * Decodes a bech32 pubkey to Amino binary, which is then decoded to a type/value object.
+ * The bech32 prefix is ignored and discareded.
+ *
+ * @param bechEncoded the bech32 encoded pubkey
+ */
+export function decodeBech32Pubkey(bechEncoded: string): PubKey {
+  const { data } = Bech32.decode(bechEncoded);
+  return decodeAminoPubkey(data);
+}
+
+/**
+ * Encodes a public key to binary Amino.
+ */
+export function encodeAminoPubkey(pubkey: PubKey): Uint8Array {
   let aminoPrefix: Uint8Array;
   switch (pubkey.type) {
     // Note: please don't add cases here without writing additional unit tests
     case pubkeyType.secp256k1:
       aminoPrefix = pubkeyAminoPrefixSecp256k1;
       break;
+    case pubkeyType.ed25519:
+      aminoPrefix = pubkeyAminoPrefixEd25519;
+      break;
     default:
       throw new Error("Unsupported pubkey type");
   }
+  return new Uint8Array([...aminoPrefix, ...fromBase64(pubkey.value)]);
+}
 
-  const data = new Uint8Array([...aminoPrefix, ...fromBase64(pubkey.value)]);
-  return Bech32.encode(prefix, data);
+/**
+ * Encodes a public key to binary Amino and then to bech32.
+ *
+ * @param pubkey the public key to encode
+ * @param prefix the bech32 prefix (human readable part)
+ */
+export function encodeBech32Pubkey(pubkey: PubKey, prefix: string): string {
+  return Bech32.encode(prefix, encodeAminoPubkey(pubkey));
 }