+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.22.1 (unreleased)
+
+- @cosmjs/cli: Import `encodeAminoPubkey`, `encodeBech32Pubkey`,
+  `decodeAminoPubkey` and `decodeBech32Pubkey` by default.
+- @cosmjs/launchpad: Add ed25519 support to `encodeBech32Pubkey`.
+- @cosmjs/launchpad: Add `encodeAminoPubkey` and `decodeAminoPubkey`.
+- @cosmjs/utils: Add `arrayContentEquals`.
+
 ## 0.22.0 (2020-08-03)
 
 - @cosmjs/cli: Now supports HTTPs URLs for `--init` code sources.